[CleanChromecast] Move all of the files around to match other sample apps.  removed files that are now part of the core SDK jar.  refactored PlayerActivity to have more clarity with valuable code
