+++ ChromecastSampleApp/app/src/main/java/com/ooyala/sample/lists/ChromecastListActivity.java
@@ -1,4 +1,4 @@
-package com.ooyala.sample.ChromecastSampleApp;
+package com.ooyala.sample.lists;
 
 import android.content.Intent;
 import android.graphics.Color;
@@ -13,11 +13,14 @@ import android.widget.AdapterView;
 import android.widget.ArrayAdapter;
 import android.widget.ListView;
 import android.widget.TextView;
+import com.ooyala.sample.utils.ChromecastPlayerSelectionOption;
+import com.ooyala.sample.players.ChromecastPlayerActivity;
+import com.ooyala.sample.R;
 
 import com.google.android.libraries.cast.companionlibrary.widgets.MiniController;
 import com.ooyala.android.castsdk.CastManager;
 
-public class ChromecastListActivity extends AppCompatActivity {
+public class ChromecastListActivity extends AppCompatActivity implements AdapterView.OnItemClickListener {
   private static final String TAG = "ChromecastListActivity";
 
   private MiniController defaultMiniController;
@@ -31,21 +34,8 @@ public class ChromecastListActivity extends AppCompatActivity {
     CastManager.getCastManager().getVideoCastManager().setStopOnDisconnect(false);
     setContentView(R.layout.start_view);
 
-    videoList = new ChromecastPlayerSelectionOption[] {
-        new ChromecastPlayerSelectionOption("HLS Asset", "Y1ZHB1ZDqfhCPjYYRbCEOz0GR8IsVRm1", "FoeG863GnBL4IhhlFC1Q2jqbkH9m", "http://www.ooyala.com"),
-        new ChromecastPlayerSelectionOption("MP4 Video", "h4aHB1ZDqV7hbmLEv4xSOx3FdUUuephx", "FoeG863GnBL4IhhlFC1Q2jqbkH9m", "http://www.ooyala.com"),
-        new ChromecastPlayerSelectionOption("VOD CC", "92cWp0ZDpDm4Q8rzHfVK6q9m6OtFP-ww", "FoeG863GnBL4IhhlFC1Q2jqbkH9m", "http://www.ooyala.com"),
-        new ChromecastPlayerSelectionOption("Encrypted HLS Asset", "ZtZmtmbjpLGohvF5zBLvDyWexJ70KsL-", "FoeG863GnBL4IhhlFC1Q2jqbkH9m", "http://www.ooyala.com"),
-        // Will play Playready Smooth on Chromecast, Clear HLS on device
-        new ChromecastPlayerSelectionOption("Playready Smooth, Clear HLS Backup", "pkMm1rdTqIAxx9DQ4-8Hyp9P_AHRe4pt", "FoeG863GnBL4IhhlFC1Q2jqbkH9m", "http://www.ooyala.com"),
-        new ChromecastPlayerSelectionOption("2 Assets autoplayed", "Y1ZHB1ZDqfhCPjYYRbCEOz0GR8IsVRm1", "92cWp0ZDpDm4Q8rzHfVK6q9m6OtFP-ww", "FoeG863GnBL4IhhlFC1Q2jqbkH9m", "http://www.ooyala.com"),
-        //These asset will not be configured correctly. To test your OPT-enabled assets, you need:
-        // 1. an OPT-enabled embed code (set here)
-        // 2. the correlating PCode (set in the PlayerViewController)
-        // 3. an API Key and Secret for the provider to locally-sign the authorization (set in the PlayerViewController)
-        new ChromecastPlayerSelectionOption("Ooyala Player Token Asset (unconfigured)", "0yMjJ2ZDosUnthiqqIM3c8Eb8Ilx5r52", "FoeG863GnBL4IhhlFC1Q2jqbkH9m", "http://www.ooyala.com"),
-        new ChromecastPlayerSelectionOption("Concurrent Streams (unconfigured)", "pwc3J0dTpAL7gMLFNVt2ks2v8j3qOKCS", "FoeG863GnBL4IhhlFC1Q2jqbkH9m", "http://www.ooyala.com"),
-    };
+    videoList = getVideoList();
+
     //Create the adapter for the ListView
     ArrayAdapter<String> selectionAdapter = new ArrayAdapter<String>(this, R.layout.list_activity_list_item);
     for(ChromecastPlayerSelectionOption video : videoList) {
@@ -53,21 +43,11 @@ public class ChromecastListActivity extends AppCompatActivity {
     }
     selectionAdapter.notifyDataSetChanged();
 
+    //Populate the listView
     ListView listView = (ListView) findViewById(R.id.listView);
     listView.setAdapter(selectionAdapter);
+    listView.setOnItemClickListener(this);
 
-    final Intent intent = new Intent(this, ChromecastPlayerActivity.class);
-    listView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
-      @Override
-      public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
-        intent.putExtra("embedcode", videoList[position].embedCode);
-        intent.putExtra("embedcode2", videoList[position].embedCode2);
-        intent.putExtra("pcode", videoList[position].pcode);
-        intent.putExtra("domain", videoList[position].domain);
-        startActivity(intent);
-      }
-    });
-    
     ActionBar actionBar = getSupportActionBar();
     actionBar.setBackgroundDrawable(new ColorDrawable(Color.TRANSPARENT));
     getSupportActionBar().setDisplayShowTitleEnabled(false);
@@ -81,6 +61,16 @@ public class ChromecastListActivity extends AppCompatActivity {
 
 // Uncomment it if you want to activate the customized sample app in our sample app
 //    customizedMiniController = (OOMiniController) findViewById(R.id.miniController2);
+  }
+
+  @Override
+  public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
+    final Intent intent = new Intent(this, ChromecastPlayerActivity.class);
+    intent.putExtra("embedcode", videoList[position].embedCode);
+    intent.putExtra("embedcode2", videoList[position].embedCode2);
+    intent.putExtra("pcode", videoList[position].pcode);
+    intent.putExtra("domain", videoList[position].domain);
+    startActivity(intent);
   }
 
   @Override
@@ -122,5 +112,26 @@ public class ChromecastListActivity extends AppCompatActivity {
     Log.d(TAG, "onPause()");
     CastManager.getCastManager().onPause();
     super.onPause();
+  }
+
+  /**
+    Generate the list of all videos for the list
+   */
+  private ChromecastPlayerSelectionOption[] getVideoList() {
+    return new ChromecastPlayerSelectionOption[] {
+      new ChromecastPlayerSelectionOption("HLS Asset", "Y1ZHB1ZDqfhCPjYYRbCEOz0GR8IsVRm1", "FoeG863GnBL4IhhlFC1Q2jqbkH9m", "http://www.ooyala.com"),
+      new ChromecastPlayerSelectionOption("MP4 Video", "h4aHB1ZDqV7hbmLEv4xSOx3FdUUuephx", "FoeG863GnBL4IhhlFC1Q2jqbkH9m", "http://www.ooyala.com"),
+      new ChromecastPlayerSelectionOption("VOD CC", "92cWp0ZDpDm4Q8rzHfVK6q9m6OtFP-ww", "FoeG863GnBL4IhhlFC1Q2jqbkH9m", "http://www.ooyala.com"),
+      new ChromecastPlayerSelectionOption("Encrypted HLS Asset", "ZtZmtmbjpLGohvF5zBLvDyWexJ70KsL-", "FoeG863GnBL4IhhlFC1Q2jqbkH9m", "http://www.ooyala.com"),
+      // Will play Playready Smooth on Chromecast, Clear HLS on device
+      new ChromecastPlayerSelectionOption("Playready Smooth, Clear HLS Backup", "pkMm1rdTqIAxx9DQ4-8Hyp9P_AHRe4pt", "FoeG863GnBL4IhhlFC1Q2jqbkH9m", "http://www.ooyala.com"),
+      new ChromecastPlayerSelectionOption("2 Assets autoplayed", "Y1ZHB1ZDqfhCPjYYRbCEOz0GR8IsVRm1", "92cWp0ZDpDm4Q8rzHfVK6q9m6OtFP-ww", "FoeG863GnBL4IhhlFC1Q2jqbkH9m", "http://www.ooyala.com"),
+      //These asset will not be configured correctly. To test your OPT-enabled assets, you need:
+      // 1. an OPT-enabled embed code (set here)
+      // 2. the correlating PCode (set in the PlayerViewController)
+      // 3. an API Key and Secret for the provider to locally-sign the authorization (set in the PlayerViewController)
+      new ChromecastPlayerSelectionOption("Ooyala Player Token Asset (unconfigured)", "0yMjJ2ZDosUnthiqqIM3c8Eb8Ilx5r52", "FoeG863GnBL4IhhlFC1Q2jqbkH9m", "http://www.ooyala.com"),
+      new ChromecastPlayerSelectionOption("Concurrent Streams (unconfigured)", "pwc3J0dTpAL7gMLFNVt2ks2v8j3qOKCS", "FoeG863GnBL4IhhlFC1Q2jqbkH9m", "http://www.ooyala.com"),
+    };
   }
 }
