+++ ChromecastSampleApp/app/src/main/java/com/ooyala/sample/utils/ChromecastPlayerSelectionOption.java
@@ -1,4 +1,4 @@
-package com.ooyala.sample.ChromecastSampleApp;
+package com.ooyala.sample.utils;
 
 public class ChromecastPlayerSelectionOption {
   public String title;
