+++ ChromecastSampleApp/app/src/main/java/com/ooyala/sample/utils/SampleApplication.java
@@ -1,4 +1,4 @@
-package com.ooyala.sample.ChromecastSampleApp;
+package com.ooyala.sample.utils;
 
 import android.app.Application;
 import android.content.pm.ApplicationInfo;
@@ -6,9 +6,7 @@ import android.content.pm.ApplicationInfo;
 import com.ooyala.android.castsdk.CastManager;
 import com.ooyala.android.castsdk.CastOptions;
 
-/**
- * Created by zchen on 10/9/15.
- */
+import com.ooyala.sample.players.ChromecastPlayerActivity;
 
 public class SampleApplication extends Application {
   private final String NAMESPACE = "urn:x-cast:ooyala";
